+++ binless/src/fast_binless.hpp
@@ -27,12 +27,13 @@ std::vector<double> remove_signal_degeneracy(const FastSignalData& data);
 std::vector<double> shift_signal(const FastSignalData& data);
 
 Rcpp::List binless(const DataFrame obs, unsigned nbins, double alpha, const NumericVector lam2, const NumericVector lam1 = NumericVector(1),
-                   unsigned nouter=25, double tol_val=2e-1, unsigned bg_steps=5, unsigned free_decay=10000);
+                   unsigned nouter=25, double tol_val=2e-1, unsigned bg_steps=5, unsigned free_decay=10000, bool compute_patchnos=true);
 
 Rcpp::List binless_eval_cv(const List obs, double alpha, const NumericVector lam2, const NumericVector lam1 = NumericVector(1),
                            unsigned group = 0, double tol_val = 1e-1);
 
-Rcpp::DataFrame binless_difference(const List obs, unsigned ref, double alpha, const NumericVector lam2, const NumericVector lam1 = NumericVector(1), double tol_val=2e-1);
+Rcpp::DataFrame binless_difference(const List obs, unsigned ref, double alpha, const NumericVector lam2,
+                                   const NumericVector lam1 = NumericVector(1), double tol_val=2e-1, bool compute_patchnos=true);
 
 #include "fast_binless.ipp"