+++ binless/src/fast_dataframe.cpp
@@ -5,7 +5,8 @@
 namespace binless {
 namespace fast {
 
-Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const ExposureEstimator& expo, const BiasEstimator& bias, const DecayEstimator& dec, const NumericVector lam1, double tol_val) {
+Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const ExposureEstimator& expo, const BiasEstimator& bias,
+                                 const DecayEstimator& dec, const NumericVector lam1, double tol_val, bool compute_patchnos) {
   //bias, decay, signal with decay and exposures, and log_background matrix (w/ offset)
   std::vector<double> biasmat,decaymat,signal,binless,background;
   biasmat.reserve(data.get_N());
@@ -33,14 +34,18 @@ Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const ExposureEst
   }
   //build patchnos
   std::vector<double> patchnos;
-  std::vector<unsigned> dbin1(data.get_bin1()), dbin2(data.get_bin2());
   patchnos.reserve(data.get_N());
-  for (unsigned dset=0; dset<data.get_ndatasets(); ++dset) {
-    std::vector<unsigned> d_bin1(dbin1.cbegin() + dset*data.get_ncells(), dbin1.cbegin() + (dset+1)*data.get_ncells());
-    std::vector<unsigned> d_bin2(dbin2.cbegin() + dset*data.get_ncells(), dbin2.cbegin() + (dset+1)*data.get_ncells());
-    std::vector<double> d_beta(log_signal.cbegin() + dset*data.get_ncells(), log_signal.cbegin() + (dset+1)*data.get_ncells());
-    IntegerVector patchno = get_patch_numbers(data.get_nbins(), tol_val, Rcpp::wrap(d_bin1), Rcpp::wrap(d_bin2), Rcpp::wrap(d_beta));
-    for (const double v : patchno) patchnos.push_back(v);
+  if (compute_patchnos) {
+    std::vector<unsigned> dbin1(data.get_bin1()), dbin2(data.get_bin2());
+    for (unsigned dset=0; dset<data.get_ndatasets(); ++dset) {
+      std::vector<unsigned> d_bin1(dbin1.cbegin() + dset*data.get_ncells(), dbin1.cbegin() + (dset+1)*data.get_ncells());
+      std::vector<unsigned> d_bin2(dbin2.cbegin() + dset*data.get_ncells(), dbin2.cbegin() + (dset+1)*data.get_ncells());
+      std::vector<double> d_beta(log_signal.cbegin() + dset*data.get_ncells(), log_signal.cbegin() + (dset+1)*data.get_ncells());
+      IntegerVector patchno = get_patch_numbers(data.get_nbins(), tol_val, Rcpp::wrap(d_bin1), Rcpp::wrap(d_bin2), Rcpp::wrap(d_beta));
+      for (const double v : patchno) patchnos.push_back(v);
+    }
+  } else {
+    patchnos.assign(data.get_N(),-1.);
   }
   //return
   return Rcpp::DataFrame::create(_["name"]=data.get_name_factor(),
@@ -62,21 +67,25 @@ Rcpp::DataFrame get_as_dataframe(const FastData<Signal>& data, const ExposureEst
 }
 
 
-Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data, const NumericVector lam1, double tol_val) {
+Rcpp::DataFrame get_as_dataframe(const FastData<Difference>& data, const NumericVector lam1, double tol_val, bool compute_patchnos) {
   std::vector<double> log_difference = data.get_log_difference();
   std::vector<double> difference;
   difference.reserve(log_difference.size());
   for (const double v : log_difference) difference.push_back(std::exp(v));
   //build patchnos
-  std::vector<unsigned> dbin1(data.get_bin1()), dbin2(data.get_bin2());
   std::vector<double> patchnos;
   patchnos.reserve(data.get_N());
-  for (unsigned dset=0; dset<data.get_ndatasets(); ++dset) {
-    std::vector<unsigned> d_bin1(dbin1.cbegin() + dset*data.get_ncells(), dbin1.cbegin() + (dset+1)*data.get_ncells());
-    std::vector<unsigned> d_bin2(dbin2.cbegin() + dset*data.get_ncells(), dbin2.cbegin() + (dset+1)*data.get_ncells());
-    std::vector<double> d_beta(log_difference.cbegin() + dset*data.get_ncells(), log_difference.cbegin() + (dset+1)*data.get_ncells());
-    IntegerVector patchno = get_patch_numbers(data.get_nbins(), tol_val, Rcpp::wrap(d_bin1), Rcpp::wrap(d_bin2), Rcpp::wrap(d_beta));
-    for (const double v : patchno) patchnos.push_back(v);
+  if (compute_patchnos) {
+    std::vector<unsigned> dbin1(data.get_bin1()), dbin2(data.get_bin2());
+    for (unsigned dset=0; dset<data.get_ndatasets(); ++dset) {
+      std::vector<unsigned> d_bin1(dbin1.cbegin() + dset*data.get_ncells(), dbin1.cbegin() + (dset+1)*data.get_ncells());
+      std::vector<unsigned> d_bin2(dbin2.cbegin() + dset*data.get_ncells(), dbin2.cbegin() + (dset+1)*data.get_ncells());
+      std::vector<double> d_beta(log_difference.cbegin() + dset*data.get_ncells(), log_difference.cbegin() + (dset+1)*data.get_ncells());
+      IntegerVector patchno = get_patch_numbers(data.get_nbins(), tol_val, Rcpp::wrap(d_bin1), Rcpp::wrap(d_bin2), Rcpp::wrap(d_beta));
+      for (const double v : patchno) patchnos.push_back(v);
+    }
+  } else {
+    patchnos.assign(data.get_N(),-1.);
   }
   return Rcpp::DataFrame::create(_["name"]=data.get_name_factor(),
                                  _["bin1"]=data.get_bin1_factor(),